+++ client/rpc/src/integration-test/kotlin/net/corda/client/rpc/CordaRPCClientTest.kt
@@ -3,7 +3,10 @@ package net.corda.client.rpc
 import net.corda.core.contracts.DOLLARS
 import net.corda.core.flows.FlowException
 import net.corda.core.getOrThrow
+import net.corda.core.messaging.FlowHandle
+import net.corda.core.messaging.FlowProgressHandle
 import net.corda.core.messaging.startFlow
+import net.corda.core.messaging.startTrackedFlow
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.random63BitValue
 import net.corda.core.serialization.OpaqueBytes
@@ -21,8 +24,7 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import java.util.*
-import kotlin.test.assertEquals
-import kotlin.test.assertTrue
+import kotlin.test.*
 
 class CordaRPCClientTest : NodeBasedTest() {
     private val rpcUser = User("user1", "test", permissions = setOf(
@@ -69,7 +71,7 @@ class CordaRPCClientTest : NodeBasedTest() {
         println("Creating proxy")
         val proxy = client.proxy()
         println("Starting flow")
-        val flowHandle = proxy.startFlow(
+        val flowHandle = proxy.startTrackedFlow(
                 ::CashIssueFlow,
                 20.DOLLARS, OpaqueBytes.of(0), node.info.legalIdentity, node.info.legalIdentity)
         println("Started flow, waiting on result")
@@ -91,6 +93,16 @@ class CordaRPCClientTest : NodeBasedTest() {
     }
 
     @Test
+    fun `check basic flow has no progress`() {
+        client.start(rpcUser.username, rpcUser.password)
+        val proxy = client.proxy()
+        proxy.startFlow(::CashPaymentFlow, 100.DOLLARS, node.info.legalIdentity).use {
+            assertFalse(it is FlowProgressHandle<*>)
+            assertTrue(it is FlowHandle<*>)
+        }
+    }
+
+    @Test
     fun `get cash balances`() {
         println("Starting client")
         client.start(rpcUser.username, rpcUser.password)
@@ -105,9 +117,7 @@ class CordaRPCClientTest : NodeBasedTest() {
                 node.info.legalIdentity, node.info.legalIdentity
         )
         println("Started issuing cash, waiting on result")
-        flowHandle.progress.subscribe {
-            println("CashIssue PROGRESS $it")
-        }
+        flowHandle.returnValue.get()
 
         val finishCash = proxy.getCashBalances()
         println("Cash Balances: $finishCash")